+++ modulo3/Astromatch/astromach2/src/components/TelaInicial/styled.js
@@ -8,7 +8,6 @@ justify-content: center;
 align-items: center;
 height: 100vh;
 background-color: #f5f5f5;
-
 `
 
 export const CardFilho = styled.div`
